+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## UNRELEASED
+- Fix to TERT protein annotation (G>A -> C>T)
+
 ## v.1.5.5: 2024-03-05
 - Further fixes to pWGS Cardea helper to allow multiple projects for one donor